# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/otp/request": {
      "post": {
        "operationId": "AuthController_otpRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_otpVerify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "auth"
        ]
      }
    },
...
    "/businesses": {
      "post": {
        "operationId": "BusinessesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      },
      "get": {
        "operationId": "BusinessesController_mine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{businessId}": {
      "get": {
        "operationId": "BusinessesController_details",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fiche de l'entreprise : ouverte à tout membre (la lecture des données métier a ses propres permissions).",
        "tags": [
          "businesses"
        ]
      },
      "patch": {
        "operationId": "BusinessesController_update",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{businessId}/activate": {
      "post": {
        "operationId": "BusinessesController_activate",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{businessId}/members": {
      "get": {
        "operationId": "BusinessesController_members",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{businessId}/members/{userId}": {
      "patch": {
        "operationId": "BusinessesController_changeRole",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      },
      "delete": {
        "operationId": "BusinessesController_remove",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{businessId}/members/{userId}/permissions": {
      "put": {
        "operationId": "BusinessesController_setPermissions",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPermissionOverridesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{businessId}/members/{userId}/status": {
      "put": {
        "operationId": "BusinessesController_setStatus",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeMemberStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{businessId}/invitations": {
      "get": {
        "operationId": "BusinessesController_invitations",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      },
      "post": {
        "operationId": "BusinessesController_invite",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/businesses/{businessId}/invitations/{invitationId}": {
      "delete": {
        "operationId": "BusinessesController_revoke",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "businesses"
        ]
      }
    },
    "/me/invitations": {
      "get": {
        "operationId": "MyInvitationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "invitations"
        ]
      }
    },
    "/me/invitations/{invitationId}/accept": {
      "post": {
        "operationId": "MyInvitationsController_accept",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "invitations"
        ]
      }
    },
    "/me/invitations/{invitationId}/decline": {
      "post": {
        "operationId": "MyInvitationsController_decline",
        "parameters": [
          {
            "name": "invitationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "invitations"
        ]
      }
    },
...
  },
  "info": {
    "title": "THY API",
    "description": "API de l'écosystème THY. Authentification : jeton d'accès (Bearer) obtenu par OTP (POST /auth/otp/request puis /auth/otp/verify). Erreurs : { error: { code, message, details?, requestId } }. Les routes « à plat » de THY Business (ex. /products) visent l'entreprise active du jeton (POST /businesses/{id}/activate).",
    "version": "1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "OtpRequestDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "phone"
        ]
      },
      "OtpVerifyDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 20,
            "maxLength": 200
          },
          "businessId": {
            "type": "string",
            "description": "Entreprise active à conserver dans le nouveau jeton d'accès (vérifiée en base).",
            "format": "uuid"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "CreateBusinessDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "businessType": {
            "type": "string",
            "maxLength": 60
          },
          "country": {
            "type": "string",
            "pattern": "^[A-Z]{2}$"
          },
          "currency": {
            "type": "string",
            "pattern": "^[A-Z]{3}$"
          },
          "timezone": {
            "type": "string",
            "maxLength": 64
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateBusinessDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "businessType": {
            "type": "string",
            "maxLength": 60
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string",
            "maxLength": 200
          }
        }
      },
      "ChangeMemberRoleDto": {
        "type": "object",
        "properties": {
          "roleCode": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "CASHIER",
              "STOCK_KEEPER",
              "ACCOUNTANT",
              "VIEWER"
            ]
          }
        },
        "required": [
          "roleCode"
        ]
      },
      "SetPermissionOverridesDto": {
        "type": "object",
        "properties": {
          "overrides": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            }
          }
        },
        "required": [
          "overrides"
        ]
      },
      "ChangeMemberStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "InviteMemberDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          },
          "roleCode": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "CASHIER",
              "STOCK_KEEPER",
              "ACCOUNTANT",
              "VIEWER"
            ]
          }
        },
        "required": [
          "phone",
          "roleCode"
        ]
      },
...
    }
  }
}